services/python-ai-engine/requirements.txt
grpcio==1.74.0
grpcio-tools==1.74.0
protobuf==6.31.1
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
langchain-google-genai==2.1.8
langchain-experimental==0.3.4
openai==1.98.0
# google-generativeai>=0.8.5
python-dotenv==1.1.1
pydantic==2.11.7
# Firebase Firestore dependencies (replacing MongoDB)
firebase-admin==6.6.0
google-cloud-firestore==2.21.0
# motor==3.7.1  # MongoDB async driver - DEPRECATED
# pymongo==4.13.2  # MongoDB driver - DEPRECATED
redis==6.2.0
celery==5.5.3
structlog==25.4.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-instrumentation-grpc==0.57b0
qdrant-client==1.15.1
beautifulsoup4==4.13.4
requests==2.32.4
duckduckgo-search==8.1.1
wikipedia==1.4.0
httpx==0.28.1
# AWS S3 for video storage
boto3==1.35.7
botocore==1.35.7
